+++ CHANGELOG.md
@@ -10,7 +10,11 @@
 
 - Отладочный атрибут `X-Debug-Rbac`;
 - Возможность указывать выходную схему аутентификации через параметр конфига `output_scheme`;
-- Changelog.
+- Поддержка схемы аутентификации [MyAuth2](https://github.com/ozzy-ext-myauth/specification/blob/master/v2/myauth-authentication-2.md);
+- Добавлен журнал изменений.
+- В [readme](./readme.md) добавлен раздел "Особенности реализации схем аутентификации"
+- В [readme](./readme.md) добавлен раздел "Отладка"
+- Отладочные заголовки `X-Debug-Claim-...`
 
 ### Изменено